infrastructure/requirements.txt
aws-cdk-lib<3
chalice<2
python-dotenv<2